* Improved handling of pthreads stack size.

+++ src/cfg.h
@@ -428,7 +428,7 @@ struct configuration {
   u_int16_t pkt_len_distrib_bins_lookup[ETHER_JUMBO_MTU+1];
   int use_ip_next_hop;
   int tmp_net_own_field;
-  int thread_stack;
+  size_t thread_stack;
 };
 
 /* prototypes */
